Move triplets to separate module
