+++ salmon/__init__.py
@@ -1,8 +1,8 @@
-from .frontend import app as frontend
-from .backend import app as backend
+from .frontend import app as app
+from .backend import app as app_algs
 
 from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
 
-frontend.version = backend.version = __version__
+app.version = app_algs.version = __version__
