+++ salmon/backend/__init__.py
@@ -1,5 +1,1 @@
 from .core import app
-from . import algs
-
-__version__ = "0.1.0"
-app.version = __version__
