+++ salmon/triplets/algs/__init__.py
