+++ salmon/triplets/algs/_round_robin.py
