+++ salmon/triplets/manager.py
