+++ salmon/utils.py
